Refactor and simplify S/MIME signing flow
Refactored the S/MIME signing process to streamline logic, improve maintainability, and reduce code duplication. There was a general logic flaw in the whole SMIME signing code. It would only work with quoted-printable and raw encodings, while base64 and others would fail. Instead of forcing to render the message body into QP, we no perform a pre-signing render run of the whole mail which will output all message parts in proper formatting and encoding. The result of this pre-sign run is then used for the signature. It currently doesn't work with multipart messages yet, but we now make use of all available go-mail tools to get the message parts. This makes a bunch of methods in the smime.go obsolete.

 Removed unused and redundant methods, added safeguards for in-progress signing, and improved boundary generation for multipart messages. Updated existing tests to align with the new implementation.