Fix regression of custom smtp.Auth not working.
This change re-introduces custom smtp.Auth support, which has been removed by accident. See #428. To make sure this does not happen again, a couple of tests have been added that make sure the custom auth is working as intended. This includes a `testCustomAuth` struct to verify the "CUSTOM" method and enhancements in the client to handle custom auth types. Resolves issue #428.