BREAKING: Refactor Reply-To header handling.
Changed Reply-To header to an AddrHeader. Since Reply-To is actually an address header and formats the mail address accordingly, this could lead to potential double encoding if the address and name contain special characters. This is potentially a breaking change, since we are changing a public type but as long as the user makes use of `msg.ReplyTo()` this should not cause any problems. This commit fixes #440.