Fix "short writes" in Base64LineBreaker.Write
The count of bytes written is incorrectly reported by
Base64LineBreaker.Write. This issue is not visible externally as
Base64LineBreaker.Write is only called by encoding/base64.Encoder.Write
(like in the original implementation in stdlib encoding/pem.lineBreaker)
which ignores the returned bytes count.

The issue is not visible either with the test suite (not even via
fuzzing) because the tests use Base64LineBreaker through the
encoding/base64.Encoder.

So we could just keep the existing code. The only risk is if
encoding/base64 behaviour changes in the future.

See also https://go.dev/cl/653675 which aims to fix the same issue in
encoding/pem.