Make Base64LineBreaker private
Rename `Base64LineBreaker` to `base64LineBreaker` and update associated methods, variables, and error messages. As Olivier Mengué pointed out in #444, making the API public was a poor choice by myself, since the struct has no public fields and since the `io.Writer` is required, nobody except go-mail would ever be able to use the API. This is a breaking change, but since nobody would be able to use the API anyways, it should not affect any user.